requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.1.1
Flask-Bcrypt==1.0.1
python-dotenv==0.19.1
Werkzeug==2.0.1
requests==2.26.0
PyPDF2==2.12.1
numpy==1.24.3
scikit-learn==1.2.2
